opencv_webapp_lecture/requirements.txt
Django==2.0.6
numpy==1.14.4
opencv-python==3.4.1.15
Pillow==5.1.0
pytz==2018.4
